Mnvim (#9)
* This commit has a new LSP configuration

The new configuration comes from nvim's update to v0.11 which deprecates
the lsp.require config in favor of vim.lsp.enable and vim.lsp.config
which allows us to get rid off nvim-lsp and mason. It's a much simpler
configuration and separates it by lsp.

* Adding my dz60 configuration for safekeeping

This commit adds my dz60 mappings for flashing using qmktools.

* This commit adds tmux and its config

This commit adds tmux to the brew file and a short config file, when
reinstalling, just need to run tpm for the config to load agai.

ok love you bye!

---------

Co-authored-by: Rodrigo Ibarra <rodrigoibarra@Rodrigos-MacBook-Pro.local>